Refactor user_repository.py to add create_or_update_user method
This commit refactors the user_repository.py file to add a new method called create_or_update_user. This method is responsible for creating or updating a user in the database based on the provided user_info dictionary. If the user already exists, their name and email will be updated. If the user does not exist, a new user will be created with the provided information. The changes include adding the necessary imports, defining the method, and implementing the logic to create or update the user.